requirements.txt
pygame
python-can
cantools
mkdocs
mkdocs-material
